# Store requested detail indexes in hashes