This improves the performance of matching and sorting `TemplateDetails` based on `Hash#[]` being faster than `Array#include?` and `Array#index` (with the tradeoff that it's slower to build `TemplateDetails::Requested`).

This isn't currently in the hot path (once template lookups are cached), however I'm planning for it to be in the future.

**Before**
```
$ be ruby benchmark_details_match.rb
Warming up --------------------------------------
            matches?   444.636k i/100ms
        sort_key_for   271.076k i/100ms
Calculating -------------------------------------
            matches?      4.391M (± 1.9%) i/s -     22.232M in   5.065029s
        sort_key_for      2.737M (± 0.8%) i/s -     13.825M in   5.051870s
```

**After**
```
$ be ruby benchmark_details_match.rb
Warming up --------------------------------------
            matches?   679.783k i/100ms
        sort_key_for   602.556k i/100ms
Calculating -------------------------------------
            matches?      6.664M (± 1.2%) i/s -     33.989M in   5.100834s
        sort_key_for      5.995M (± 0.5%) i/s -     30.128M in   5.025551s
```

**Benchmark**

``` ruby
require "benchmark/ips"
require "action_view"

default_details = {
  locale: [:en],
  formats: [:html, :text, :js, :css, :ics, :csv, :vcf, :vtt, :png, :jpeg, :gif, :bmp, :tiff, :svg, :mpeg, :mp3, :ogg, :m4a, :webm, :mp4, :otf, :ttf, :woff, :woff2, :xml, :rss, :atom, :yaml, :multipart_form, :url_encoded_form, :json, :pdf, :zip, :gzip],
  variants: [],
  handlers: [:raw, :erb, :html, :builder, :ruby]
}

requested = ActionView::TemplateDetails::Requested.new(**default_details)
details = ActionView::TemplateDetails.new(nil, :erb, :html, nil)

result = Benchmark.ips do |x|
  x.config(:time => 5, :warmup => 2)

  x.report("matches?") do
    details.matches?(requested)
  end

  x.report("sort_key_for") do
    details.sort_key_for(requested)
  end
end
```